Update 03-01.md
Originally, the hint for the Interact Hint part says that there's a `Projects` function that can be used to help get the screen space position, but I feel like it's more accurate to call this a `UGameplayStatics` function as this is the namespace it is under. I assumed that `Project` was a reference to the namespace and searched accordingly, although it just ended up being a part of the function name (also assuming that I found the correct one).

It took me a lot of time to find the correct function (and to realize that it was correct, since I was searching for something different), so I feel that this change will prevent some time wasted on VM hours.